requirements.txt
torch
numpy
scikit-image
xgboost
pillow
opencv-python
scipy
torchmetrics
tqdm

